table component in separate file on admin
