--- src/app/admin/products/utils.jsx
@@ -1,84 +1,0 @@
-export function capitalize(s) {
-    return s ? s.charAt(0).toUpperCase() + s.slice(1).toLowerCase() : ''
-}
-
-export const PlusIcon = ({ size = 24, width, height, ...props }) => {
-    return (
-        <svg aria-hidden='true' fill='none' focusable='false' height={size || height} role='presentation' viewBox='0 0 24 24' width={size || width} {...props}>
-            <g fill='none' stroke='currentColor' strokeLinecap='round' strokeLinejoin='round' strokeWidth={1.5}>
-                <path d='M6 12h12' />
-                <path d='M12 18V6' />
-            </g>
-        </svg>
-    )
-}
-
-export const VerticalDotsIcon = ({ size = 24, width, height, ...props }) => {
-    return (
-        <svg aria-hidden='true' fill='none' focusable='false' height={size || height} role='presentation' viewBox='0 0 24 24' width={size || width} {...props}>
-            <path
-                d='M12 10c-1.1 0-2 .9-2 2s.9 2 2 2 2-.9 2-2-.9-2-2-2zm0-6c-1.1 0-2 .9-2 2s.9 2 2 2 2-.9 2-2-.9-2-2-2zm0 12c-1.1 0-2 .9-2 2s.9 2 2 2 2-.9 2-2-.9-2-2-2z'
-                fill='currentColor'
-            />
-        </svg>
-    )
-}
-
-export const SearchIcon = props => {
-    return (
-        <svg aria-hidden='true' fill='none' focusable='false' height='1em' role='presentation' viewBox='0 0 24 24' width='1em' {...props}>
-            <path
-                d='M11.5 21C16.7467 21 21 16.7467 21 11.5C21 6.25329 16.7467 2 11.5 2C6.25329 2 2 6.25329 2 11.5C2 16.7467 6.25329 21 11.5 21Z'
-                stroke='currentColor'
-                strokeLinecap='round'
-                strokeLinejoin='round'
-                strokeWidth='2'
-            />
-            <path d='M22 22L20 20' stroke='currentColor' strokeLinecap='round' strokeLinejoin='round' strokeWidth='2' />
-        </svg>
-    )
-}
-
-export const ChevronDownIcon = ({ strokeWidth = 1.5, ...otherProps }) => {
-    return (
-        <svg aria-hidden='true' fill='none' focusable='false' height='1em' role='presentation' viewBox='0 0 24 24' width='1em' {...otherProps}>
-            <path
-                d='m19.92 8.95-6.52 6.52c-.77.77-2.03.77-2.8 0L4.08 8.95'
-                stroke='currentColor'
-                strokeLinecap='round'
-                strokeLinejoin='round'
-                strokeMiterlimit={10}
-                strokeWidth={strokeWidth}
-            />
-        </svg>
-    )
-}
-
-export const columns = [
-    { name: 'ID', uid: 'id', sortable: true },
-    { name: 'NAME', uid: 'name', sortable: true },
-    { name: 'AGE', uid: 'age', sortable: true },
-    { name: 'ROLE', uid: 'role', sortable: true },
-    { name: 'TEAM', uid: 'team' },
-    { name: 'EMAIL', uid: 'email' },
-    { name: 'STATUS', uid: 'status', sortable: true },
-    { name: 'ACTIONS', uid: 'actions' },
-]
-
-export const statusOptions = [
-    { name: 'Active', uid: 'active' },
-    { name: 'Paused', uid: 'paused' },
-    { name: 'Vacation', uid: 'vacation' },
-]
-
-export const users = [
-    // ... your existing users array ...
-]
-
-export const statusColorMap = {
-    active: 'success',
-    paused: 'danger',
-    vacation: 'warning',
-}
-
-export const INITIAL_VISIBLE_COLUMNS = ['name', 'role', 'status', 'actions']
